requirements.txt
fastapi[standard]
pydantic
camelot-py
opencv-python-headless
ghostscript